pyproject.toml
[project]
name = "mcp-fun"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.181.0",
    "mcp[cli]>=1.14.0",
    "oauth2client>=4.1.3",
    "python-dotenv>=1.1.1",
]
